+++ class/skill.lua
@@ -3,6 +3,10 @@
 Skill = class_refs["Skill"]
 
 local callbacks = {}
+local other_callbacks = {
+    "onPreStep",
+    "onPostStep"
+}
 
 local achievement_map = {}
 
@@ -87,40 +91,41 @@ end
 methods_skill = {
 
     add_callback = function(self, callback, func)
+        local callback_id = nil
+        if      callback == "onCanActivate" then callback_id = self.on_can_activate
+        elseif  callback == "onActivate" then callback_id = self.on_activate
+        elseif  callback == "onEquipped" then callback_id = self.on_equipped
+        elseif  callback == "onUnequipped" then callback_id = self.on_unequipped
+        end
 
-        if callback == "onCanActivate" then 
-            local callback_id = self.on_can_activate
+        if callback_id then
             if not callbacks[callback_id] then callbacks[callback_id] = {} end
             table.insert(callbacks[callback_id], func)
+        
+        elseif Helper.table_has(other_callbacks, callback) then
+            if not callbacks[callback] then callbacks[callback] = {} end
+            table.insert(callbacks[callback], {self.value, func})
+
+        else log.error("Invalid callback name", 2)
         end
-        if callback == "onActivate" then 
-            local callback_id = self.on_activate
-            if not callbacks[callback_id] then callbacks[callback_id] = {} end
-            table.insert(callbacks[callback_id], func)
-        end
-        if callback == "onStep" then
-            local callback_id = self.on_step
-            if not callbacks[callback_id] then callbacks[callback_id] = {} end
-            table.insert(callbacks[callback_id], func)
-        end
-        if callback == "onEquipped" then
-            local callback_id = self.on_equipped
-            if not callbacks[callback_id] then callbacks[callback_id] = {} end
-            table.insert(callbacks[callback_id], func)
-        end
-        if callback == "onUnequipped" then
-            local callback_id = self.on_unequipped
-            if not callbacks[callback_id] then callbacks[callback_id] = {} end
-            table.insert(callbacks[callback_id], func)
-        end 
     end,
 
     clear_callbacks = function(self)
         callbacks[self.on_can_activate] = nil
         callbacks[self.on_activate] = nil
-        callbacks[self.on_step] = nil
         callbacks[self.on_equipped] = nil
         callbacks[self.on_unequipped] = nil
+        
+        for _, c in ipairs(other_callbacks) do
+            local c_table = callbacks[c]
+            if c_table then
+                for i, v in ipairs(c_table) do
+                    if v[1] == self.value then
+                        table.remove(c_table, i)
+                    end
+                end
+            end
+        end
     end,
 
     set_skill = function(self, name, desc, sprite, subimage, damage, cooldown)
@@ -219,7 +224,8 @@ methods_skill = {
     -- Callbacks
     onCanActivate   = function(self, func) self:add_callback("onCanActivate", func) end,
     onActivate      = function(self, func) self:add_callback("onActivate", func) end,
-    onStep          = function(self, func) self:add_callback("onStep", func) end,
+    onPreStep       = function(self, func) self:add_callback("onPreStep", func) end,
+    onPostStep      = function(self, func) self:add_callback("onPostStep", func) end,
     onEquipped      = function(self, func) self:add_callback("onEquipped", func) end,
     onUnequipped    = function(self, func) self:add_callback("onUnequipped", func) end
     
@@ -255,11 +261,76 @@ metatable_class["Skill"] = {
 
 gm.post_script_hook(gm.constants.callback_execute, function(self, other, result, args)
     if callbacks[args[1].value] then
+        Helper.log_hook(self, other, result, args)
+
         for _, fn in pairs(callbacks[args[1].value]) do
-            fn(Instance.wrap(args[2].value), Skill.wrap(args[3].value), args[4].value) --(actor, skill, index)
+            fn(Instance.wrap(args[2].value), args[3].value, args[4].value) --(actor, ActorSkill struct, index)
         end
     end
 end)
+
+
+
+-- ========== Callbacks ==========
+
+-- These currently only apply to Player skills to prevent lag
+-- Should be fine until someone complains about it
+
+local function skill_onPreStep(self, other, result, args)
+    if gm.variable_global_get("pause") then return end
+    
+    if callbacks["onPreStep"] then
+        local actors = Instance.find_all(gm.constants.oP)
+        for n, actor in ipairs(actors) do
+
+            -- Loop through active skills
+            for slot = 0, 3 do
+                local active_skill = actor:get_active_skill(slot)
+
+                -- Loop through callbacks
+                for _, c in ipairs(callbacks["onPreStep"]) do
+                    if c[1] == active_skill.skill_id then
+                        c[2](actor, active_skill, slot)  -- Actor, ActorSkill struct, index
+                    end
+                end
+            end
+
+        end
+    end
+end
+
+
+local function skill_onPostStep(self, other, result, args)
+    if gm.variable_global_get("pause") then return end
+    
+    if callbacks["onPostStep"] then
+        local actors = Instance.find_all(gm.constants.oP)
+        for n, actor in ipairs(actors) do
+
+            -- Loop through active skills
+            for slot = 0, 3 do
+                local active_skill = actor:get_active_skill(slot)
+
+                -- Loop through callbacks
+                for _, c in ipairs(callbacks["onPostStep"]) do
+                    if c[1] == active_skill.skill_id then
+                        c[2](actor, active_skill, slot)  -- Actor, ActorSkill struct, index
+                    end
+                end
+            end
+
+        end
+    end
+end
+
+
+
+-- ========== Initialize ==========
+
+initialize_skill = function()
+    Callback.add("preStep", "RMT-skill_onPreStep", skill_onPreStep)
+    Callback.add("postStep", "RMT-skill_onPostStep", skill_onPostStep)
+end