+++ main.lua
@@ -64,4 +64,5 @@ function __initialize()
     initialize_equipment()
     initialize_interactable()
     initialize_item()
+    initialize_skill()
 end