+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 * Helper
     * `log_hook` now logs struct variables if self and other are structs.
     * `log_struct` now adds borders above and below it.
+* Actor
+    * Added `set_default_skill`.
+* Skill
+    * Fixed onStep callback (by making a custom one; now replaced with onPreStep and onPostStep).
+    * Fixed arg2 being incorrectly wrapped as a Skill object when it is a struct.
 
 ### v1.1.16
 * Resources : `sprite_load` now returns the existing sprite of a given namespace-identifier if it already exists.